Package Update
Updates to package code. Most notably, heatmap functions now have an option to not show the annotations for rows and columns if desired and options to add a border color and to change the NA color have been added. Functions for updating annotations have been improved so that the user may add annotations for a portion of the data and the stored annotations will be properly updated within the function.